requirements.txt
configparser==5.0.0
pylint==2.5.3
pytest==5.4.3
pytest-mock==3.2.0
pytest-cov==2.10.0
GitPython==3.1.3